@@ -1,22 +1,24 @@
 package hajiboot;
 
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
-import org.springframework.web.client.RestTemplate;
+import org.springframework.web.client.RestClient;
 
 @Component
-@ConditionalOnProperty(name = "hajiboot3.markdown.type", havingValue = "github") // (1)
 public class GitHubApiMarkdownRenderer implements MarkdownRenderer {
-	private final RestTemplate restTemplate;
 
-	public GitHubApiMarkdownRenderer(RestTemplateBuilder restTemplateBuilder) {
-		this.restTemplate = restTemplateBuilder.build();
+	private final RestClient restClient;
+
+	public GitHubApiMarkdownRenderer(RestClient.Builder restClientBuilder) {
+		this.restClient = restClientBuilder.build();
 	}
 
 	@Override
 	public String render(String markdown) {
-		return this.restTemplate.postForObject("https://api.github.com/markdown/raw",
-				markdown, String.class);
+		return this.restClient.post()
+			.uri("https://api.github.com/markdown/raw")
+			.body(markdown)
+			.retrieve()
+			.body(String.class);
 	}
+
 }
